+++ CHANGELOG.md
@@ -3,6 +3,11 @@
 ## Upcoming...
 
 - ... <!-- Add new lines here. Version number will be decided later -->
+
+## 0.3.7
+
+_2020-09-14_
+
 - feat: Add `render` prop to `Modal` providing a `closeModal` dispatcher
 - feat: Add `aria-controls` prop on `CarouselStepper` and `CarouselPaging`
 - feat: Set `title` attr on `CarouselStepper` and `CarouselPaging` buttons